boloo/requirement.txt
Django==2.0
Python==3.8.0
djangorestframework==3.11.0
mysqlclient==1.4.6
pytz==2019.3
